feat(snap-valve): declarative thanos levels — pre-announced per-level entry divisor
Vault owner declares a future level a thanos level (setThanosLevel): every
entry drained for levels at or past the target divides by 2^shift. A
declaration lands at least 6 levels ahead — strictly before the target's
first materialization (the far-future promotion at the transition to
target-5) — so no existing ticket is ever touched and each level's entries
share one constant exponent regardless of when they were bought or drained.
Uniform division cancels in the pot-share fraction, so any declaration
(raise or un-thanos lower) is a pure pre-announced price change, EV-neutral
per raw entry. Raw units stay the external denomination everywhere: salvage
faces, owed views, and sell debits are untouched because a raw entry's
replacement cost and pot share are snap-invariant.

Bounds: shift <= SNAP_SHIFT_MAX (8); a non-zero shift must leave the
target's projected entries (previous level's FINAL pool target at the
target's price — settled history whose x00 futurePool/3 rewrite has already
run, so the basis can never decrease post-declaration) at or above
SNAP_FLOOR_ENTRIES (40M = 10M whole tickets), keeping snapping undeclarable
below runaway scale; a pending declaration locks once its 6-level window
opens and folds into the active shift at its level commit, freeing the slot
for the next declaration.

Mechanics: snapShift (active) + snapLevel/snapPendingShift (pending) pack
into the ticketCursor slot; _snapShiftFor(target) resolves the exponent in
one warm SLOAD + two compares. The drains divide each owed balance on first
touch via _snapOwedPacked, folding the shifted-out fraction through the
QTY_SCALE remainder; entriesOwedPacked widens uint40 -> uint48 with bit 40
as a snap-done marker so a budget-split resume never divides a partially
consumed balance twice; fully drained balances write back zero, preserving
the queue-membership invariant.

processFutureTicketBatch's formerly open-coded per-entry loop now routes
through _processOneTicketEntry — the single per-entry engine shared with
processTicketBatch — bringing DegenerusGameMintModule back under the
EIP-170 runtime limit (it had crossed it) and leaving one drain loop to
audit instead of two duplicates.

Foil packs keep all four lines, sixteen board entries, and the full match
game on a thanos level; the price carries the exponent instead
(10 ticket prices << s), with price-equivalent mint units scaled to match —
the same effective cost-per-entry scaling the ticket path gets from
quantity division, preserving the designed foil/ticket board-efficiency
ratio at every depth. The purchase facade shares the identical shifted
quote, so the fresh-ETH cap covers the module's charge on thanos levels.

Test: SnapValve.t.sol drives BOTH real drains — fold-math fuzz, dormant
identity, exact division, whale multi-batch single-division with mid-drain
marker, far-future-key path, declaration and un-thanos boundaries,
remainder folding bounds.

+++ test/fuzz/QueueDoubleBuffer.t.sol
@@ -42,7 +42,7 @@ contract QueueHarness is DegenerusGameAdvanceModule {
         return ticketQueue[key][idx];
     }
 
-    function getTicketsOwedPacked(uint24 key, address buyer) external view returns (uint40) {
+    function getTicketsOwedPacked(uint24 key, address buyer) external view returns (uint48) {
         return entriesOwedPacked[key][buyer];
     }
